test(core): split core api tests
